hof/tui: refactor eval module, create flex-panel
